Refactor ReportPortalUser.ProjectDetailsMapper to GroupProjectDetailsRecord.
